+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [1.0.2] - 24-10-2025
+
+### Added
+- Import account script
+- Import permission scripts
+
 ## [1.0.1] - 01-07-2025
 
 ### Added